requirements.txt
# Core
streamlit>=1.30.0
pandas>=2.0.0
numpy>=1.24.0

# Visualization
plotly>=5.18.0
matplotlib>=3.7.0

# LLM & NLP
transformers>=4.36.0
torch>=2.1.0
accelerate>=0.25.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-huggingface>=0.0.1

# Data Processing
openpyxl>=3.1.0
chardet>=5.2.0
scipy>=1.11.0

# Utilities
python-dotenv>=1.0.0
